Several improvements, cleanup, and implementation of async engine.
