+++ alembic.ini
@@ -52,7 +52,7 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = postgresql+psycopg2://db_user:db_password@db_host/db_name
+#sqlalchemy.url = postgresql+asyncpg://user:pass@host/db
 
 
 [post_write_hooks]
