+++ settings.py
@@ -5,4 +5,4 @@ host = config.get('db_host')
 database = config.get('db_name')
 user = config.get('db_user')
 password = config.get('db_password')
-connection_string = f"postgresql+psycopg2://{user}:{password}@{host}/{database}"
+connection_string = f"postgresql+asyncpg://{user}:{password}@{host}/{database}"
